# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel management API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/room-guests/add-rooms": {
            "post": {
                "tags": [
                    "RoomGuest"
                ],
                "summary": "Add rooms to a guest",
                "description": "Add rooms to a guest",
                "operationId": "04814971e3bdb3c0b00898ad3dfd5b5b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "guest_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/room-guests/add-guests": {
            "post": {
                "tags": [
                    "RoomGuest"
                ],
                "summary": "Add guests to a room",
                "description": "Add guests to a room",
                "operationId": "de5df8dd22c7b169ecf0ee9ed7b8a941",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "guest_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/room-guest/{id}": {
            "get": {
                "tags": [
                    "RoomGuest"
                ],
                "summary": "Get a specific room-guest",
                "description": "Retrieve a specific roomguest by ID",
                "operationId": "959088b547dfad31b5bc2cd83d6e5112",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room-guest",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Room-Guest not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "RoomGuest"
                ],
                "summary": "Delete a specific room-guest",
                "description": "Delete a specific room-guest by ID",
                "operationId": "8199cecb49dfa24ce683efd02ce64095",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the room-guest",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Room-Guest deleted successfully"
                    },
                    "404": {
                        "description": "Room-Guest not found"
                    }
                }
            }
        },
        "/api/room-guests/update-rooms": {
            "put": {
                "tags": [
                    "RoomGuest"
                ],
                "summary": "Update rooms for a guest",
                "description": "Update rooms for a guest",
                "operationId": "185e245cb03f21ac01da16a1f3381d78",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "guest_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "room_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/room-guests/update-guests": {
            "put": {
                "tags": [
                    "RoomGuest"
                ],
                "summary": "Update guests for a room",
                "description": "Update guests for a room",
                "operationId": "14af350a8759ac6cd0e9753049545c4d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "room_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "guest_ids": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    },
...
}